fix(phpfpm): Health checks fails when env is large
When the total environment is too large, such as
over 8KB, FastCGI closes the connection with
no output or log messages. cgi-fcgi returns with
an exit code of 104 and no output.

This happens even when clear_env is set in the
PHP-FPM config, since the problem seems to happen
before that would be applied.

To fix this, we prefix the cgi-fcgi command with
`env -i` to clear the environment.